Convert Bread Crumbs to Typescript
